Make create user organisation dropdown searchable
Trello: https://trello.com/c/aQiPdsiP

This uses the accessible-autocomplete component [1] to progressively
enhance [2] the existing organisation dropdown so that users with
JavaScript enabled can type into an input to search for the organisation
name.

The extra styles in `_autocomplete.scss` are intended to make the
component fit into the GOV.UK Design System and were provided by our
designer, Calum Ryan.

The organisation dropdown is not a required field and the
accessible-autocomplete component doesn't provide a way to "clear" the
selection, so I've added a "None" option. Unfortunately if I make that
the default option (instead of a blank option) then when the user first
clicks on the component, they only see organisations matching "None" in
the "search results" for the component. To address the latter I've still
included the default blank option, which never shows up in the "search
results" for the component. I then map the "None" option to the blank
option in `CreateNewUserRequest#organisation`, so that the two options
are effectively equivalent.

Note that we did consider re-using the `autocomplete` component [3] from
Whitehall, but that only seemed to allow *multiple* selections. We also
considered re-using the `select_with_search` component [4] from
Whitehall, but that relies on the choices.js library [5] and we weren't
sure that it had been as rigorously tested for accessibility.

[1]: https://github.com/alphagov/accessible-autocomplete
[2]: https://github.com/alphagov/accessible-autocomplete#progressive-enhancement
[3]: https://github.com/alphagov/whitehall/blob/0c72a9cbf768b4be8d312447b3b984007ed6fbad/app/views/components/_autocomplete.html.erb
[4]: https://github.com/alphagov/whitehall/blob/0c72a9cbf768b4be8d312447b3b984007ed6fbad/app/views/components/_select_with_search.html.erb
[5]: https://choices-js.github.io/Choices/