Add conditional global error alert to design system layout
Strictly speaking this isn't needed at the moment. However, since some
other controllers *do* set `flash[:alert]`, it makes sense to add it to
the new design system layout now.

This seems especially sensible since we're using
`GOVUKDesignSystemFormBuilder::Builder#govuk_error_summary` to generate
the error summary at the top of the page for the
`CreateNewUserRequestsController#new` page and so we have to disable the
default validation messages alert in `RequestsController#create`. If
someone added the global alert to the design system layout later, they
could easily miss the need to do this.