Fail fast if translation not found
The `I18n.t` method returns a string like "Translation missing:
en.<key>" if the translation isn't found, while the `t!` version raises
an exception. I think raising an exception is preferable as it avoids
the possibility that a test passes because we've incorrectly referenced
a missing key in both our test and production code.