Fix error message when email is invalid
I happened to notice this in a Rails console so have added tests to
avoid a regression in future.

This table includes the keys to use in our locale files[1].

I attemped to use the `with_message` qualifier for the `allow_value`
matcher[2] but couldn't get it to work as I expected.

[1]: https://guides.rubyonrails.org/v5.0.6/i18n.html#error-message-interpolation
[2]: https://matchers.shoulda.io/docs/v6.1.0/Shoulda/Matchers/ActiveModel.html#allow_value-instance_method