Convert homepage to use GOV.UK Design System
This should pave the way for more of the Support app to be moved to the
Design System.

As agreed with @jlewr the "Make new request" dropdown hasn't been ported
to the navigation in the new Design System layout because we don't think
it adds any value over the list of links that appear on the homepage.
The dropdown remains in the existing layout so that users can still jump
between forms directly rather than going via the homepage. We're hoping
that reducing the number of forms in the Support app, and improving
those that remain, will mean we won't have to implement the dropdown as
we convert more pages to the Design System because the forms will do a
better job of routing users to the right place.

Other notable changes:

- Setting `$govuk-page-width: 1140px` matches the width in Signon[1].

- The style is copied from the list of applications displayed on the
  Signon dashboard[2] but uses the `govuk-spacing` helper[3] instead
  of hardcoding 20px, and `$govuk-border-color` instead of `#cccccc`.

- The `list-group.disabled` style was made redundant in this commit and
  has been deleted.

- Each link is now a `h2` instead of `h4` given that there are no other
  `h2` or `h3` on the page.

- Some specs have been updated to explicitly target the "Feedback
  explorer" link in the navigation to retain the previous behaviour. The
  form links on the homepage previously included the label ("Feedback
  explorer") and description ("GOV.UK Anonymous Feedback") which meant
  that Capybara didn't find a match when looking for a link of "Feedback
  explorer". Now that the form links on the homepage only include the
  label ("Feedback explorer") Capybara was finding two matching elements
  so we have to specify which one we want.

- The forms a user doesn't have access to are hidden behind a Details
  component[4] in order to reduce the noise on the page.

[1]: https://github.com/alphagov/signon/blob/5530c8e54d329947b062b76e92051090b2e12e9f/app/assets/stylesheets/application.scss#L1
[2]: https://github.com/alphagov/signon/blob/5530c8e54d329947b062b76e92051090b2e12e9f/app/assets/stylesheets/application.scss#L20
[3]: https://design-system.service.gov.uk/styles/spacing/#using-the-static-spacing-helper
[4]: https://design-system.service.gov.uk/components/details/