Move options into `TechnicalFaultReport`
We're going to refactor out the `UserFacingComponents` class
altogether: it's essentially a data holder, and if we're following
the convention of the 'ContentDataFeedback' class, we're better
off holding that data closer to where it's used (i.e. here, in
'TechnicalFaultReport').

Moving the data here makes it easier to remove the
'UserFacingComponents' class in a later commit.