Refactor out the "Do not know" special case
This was being handled as an edge case in code, and its definition
(key and value) were being duplicated in several places. But it
isn't being used any differently by the application, so can simply
be added to the standard array.

I note that it was implemented as an edge case in #970 because
"it felt wrong to add [to the list] as it's not a user facing
service". I see the argument for that, but we also have a
pre-existing "GOV.UK: content" option which isn't a publishing app
and which is not treated any differently to the other options.
I think the underlying issue was that we'd named the container class
`UserFacingComponents`, which makes the addition of non-user-facing
components questionable.

As it's no longer an edge case, we can also remove the tests that
were specifically for the 'do not know' option.