Include fault context in the Zendesk subject
Our 2nd line queue is famously not-useful as it often contains a
large number of tickets called "Technical fault report". Including
the fault context helps to distinguish between tickets at a glance
and also to get a feel for which app may be causing the most
problems on a given day. The subject will now look something like
"Technical fault with Whitehall".