Move Zendesk subject string creation into 'TechnicalFaultReport'
This is about to get more complex as we'd like to handle the
"Do not know" case with a custom subject. That level of
manipulation should not happen in the 'TechnicalFaultReportTicket'
class, it should be closer to the data where we define the
different options. This will also make it far easier to test.