Set custom Zendesk ticket subject if "Do not know" was selected
We don't want to return "Technical fault with Do not know" as the
subject, so need to add a special rule.

The duplication of `do_not_know` in this class isn't ideal, but
a refactor would require changing the OPTIONS `key => value`
structure to something like `key => { name:, subject }`, and would
require making changes in multiple files. For now it feels cleaner
to have the duplication, given this is the only override we
anticipate having.