Document the interplay between the fault context and Zendesk
We recently discovered that whilst the Technical Fault Report form
submitted tickets with a tag of the form
`fault_with_{snake_cased_option_name}`, e.g. `fault_with_maslow`,
[only a subset of tags](https://govuk.zendesk.com/admin/objects-rules/rules/triggers/35985647)
routed to GOV.UK Technical 2nd Line. The rest, by default, went to
GOV.UK User Support to triage.

In other words, the only technical fault reports that came straight
to 2nd line were reports concerning the following:

- Whitehall Publisher
- Mainstream Publisher
- Travel Advice Publisher
- Specialist Publisher
- Content Publisher
- data.gov.uk

I believe this was an oversight, and probably happened due to
assumptions by contributors to this repo that all technical fault
reports go straight to 2nd line, which was not the case.

That _is_ now the case, as we've changed the trigger rule to triage
all tickets with the `technical_fault` tag, to 2nd line. However,
we're going to be transitioning to a new support model whereby
tickets are triaged directly to the teams that own the apps, and
thus we risk getting into the same kind of sticky situation with
regards to Zendesk trigger rules being out of sync with the config
in Support.

This inline comment should make contributors think twice before
making edits.