Add "Smart Answers" option
This is a bit of a change of precedent, insofar as Smart Answers
is not a publishing app.

It's not a publishing app, but it holds a lot of content, and has
been subject to a number of bug reports in the past couple of years
as it is pretty complex. Example:
https://govuk.zendesk.com/agent/tickets/2144831

It seems like having a "Smart Answers" option is more suitable
than a "GOV.UK: content" option for cases such as these.