Add 'Support app' option
The Support app is in the same tools suite as other publishing apps
so should probably be included in this list (albeit somewhat meta!).
I've added an ' app' suffix because otherwise people may interpret
the option as "I need support", rather than "there is a problem
with the Support app".