Add "Email alerts" option
We've seen examples of people using the "GOV.UK: content" option
to raise issues about emails not working:
https://govuk.zendesk.com/agent/tickets/3396413

We're hoping to triage relevant support tickets directly to the
teams that own the respective app. So having 'email alerts' as an
option enables us to triage to owners of Email Alert API, for
example, rather than having the ticket come to 2nd Line via
the "GOV.UK: content" option.