Format the error messages for the create user form
The app currently uses the default `"%{attribute} %{message}"` format for error
messages. In the create user form, we want the freedom to be able to write
error messages in the GOV.UK style rather than having to stick to the default
format.

Update the format for the create new request form only.