Setup Google Tag Manager
This allows us to implement basic GA4 tracking on all pages of the
app. It is configured using the following environment variables:
- GOOGLE_TAG_MANAGER_ID
- GOOGLE_TAG_MANAGER_AUTH (integration only)
- GOOGLE_TAG_MANAGER_PREVIEW (integration only)

If only the `GOOGLE_TAG_MANAGER_ID` is set then the default
environment of the GTM container will be used.

If `GOOGLE_TAG_MANAGER_AUTH` and `GOOGLE_TAG_MANAGER_PREVIEW` are
set then the specified GTM container environment will be used -
specifically for testing the configuration in Integration before
rolling out to Production.