Breaking up the Campaign Request Form Date Inputs
Date picker arrows in Support app stopped working after the migration to Dart Sass. It’s caused by a known issue in the way Dart Sass handles the icon as Unicode - sass/sass#1395 we were not able to find a workaround. Date picker are not very accessible. When users are filling the dates in the form they normally know them already (and don’t need to pick it by looking at a calendar).

Users see a  symbol instead of an arrow. There’s alt text (Next, Prev). It’s also possible to enter date manually.

This will allow us to drop dependency on jquery-ui-rails.

Due to delivery priorities, we're not converting the entire form into Design System at the moment.