Specify numeric input mode for date fields
inputmode="numeric" is an HTML attribute that provides a hint to devices
(mainly mobile phones and tablets) to display a numeric keyboard instead of a
standard keyboard. Even though the input type is text (allowing any text
characters), inputmode="numeric" tells the browser that the field is expected
to receive numeric input.

type="text" is the default value for the <input> element, so it’s not
explicitly defined