Rename keys access_to_whitehall_publisher -> whitehall_training
to reflect changed context of the input field.  The question changes from "Does
the user need access to Whitehall Publisher?" to "Does the user need training
to access Whitehall Publisher?"

This is consistent with custom fields labels in Zendesk. At the early stages
"[Publishing]" was used it was likely to get confusing with our Publishing
teams existing and using Zendesk so it (and the associated tags) were changed
to [Whitehall training].