Add new or existing user question to request form
This form is to be changed so that it can be used to request training
for existing users, not just new ones. This adds a new field to specify
whether the user is new or existing. I'm not changing the names of the
classes/files to reflect the extended scope of the form, at least not
yet...