Use Design System in change user form
This is the "Change an existing user's account" form

The hidden `use_design_system` field is there so that if the submitted
request is invalid, the controller can work out whether to use the
Design System from the `#create` action

The format provided in `config/locales/en.yml` tells
`error.full_message` to use just the custom message for errors on this
form. Without this override, it would default to prepending the field
name to the custom message. Using `#full_message` rather than just
`#message` in the `ErrorsHelper` module is useful for if we forget to
define a custom error message so that we get messages like "Title can't
be blank" rather than "can't be blank", but ideally we'd always provide
a user-friendly message like we do here

We have coverage for this form in a dedicated feature test, so we can be
confident that everything still works as expected