Support success alerts via flash in Design System
We already support flashed messages in the non-Design System template.
This is using a different key, but it aligns with Signon and one of the
govuk_publishing_components components. The non-Design System flow
redirects to an acknowledge path after successful request submission;
for the Design System flow we'll instead redirect to the root path and
use this flash to report success